refactor(sagas): refactor addToContract to be part of contractSaga
