+++ src/sagas/contractSaga.js
@@ -1,9 +1,17 @@
 import {
-  takeLatest, select, put
+  takeLatest, select, put, call, takeEvery
 } from 'redux-saga/effects';
+import {
+  PluginManager, List, FromMarkdown, ToMarkdown
+} from '@accordproject/markdown-editor';
+import ClausePlugin from '@accordproject/cicero-ui/dist/plugins/ClausePlugin';
+import { Value } from 'slate';
+import uuidv4 from 'uuidv4';
+
 import * as actions from '../actions/contractActions';
 import * as appActions from '../actions/appActions';
 import * as contractSelectors from '../selectors/contractSelectors';
+import { addTemplateObjectToStore } from './templatesSaga';
 
 /**
  * saga to update the contract in the store if it has changed
@@ -19,6 +27,34 @@ export function* updateDocument(action) {
   }
 }
 
+/**
+ * saga which adds a clause node to the current slate value
+ */
+export function* addToContract(action) {
+  const pluginManager = new PluginManager([List(), ClausePlugin()]);
+  const fromMarkdown = new FromMarkdown(pluginManager);
+  const toMarkdown = new ToMarkdown(pluginManager);
+
+  const templateObj = yield call(addTemplateObjectToStore, action);
+
+  const slateValue = yield select(contractSelectors.slateValue);
+  const { metadata } = templateObj;
+
+  const currentPosition = slateValue.selection.anchor.path.get(0);
+  const clauseMd = `\`\`\` <clause src=${action.uri} clauseId=${uuidv4()}>
+  ${metadata.getSample()}
+  \`\`\``;
+  const value = fromMarkdown.convert(clauseMd);
+  const clauseNode = value.toJSON().document.nodes[0];
+
+  const newSlateValue = JSON.parse(JSON.stringify(slateValue.toJSON()));
+  const newMd = toMarkdown.convert(newSlateValue);
+  const { nodes } = newSlateValue.document;
+  nodes.splice(currentPosition, 0, clauseNode);
+  yield put(actions.documentEdited(Value.fromJSON(newSlateValue), newMd));
+}
+
 export const contractSaga = [
   takeLatest('DOCUMENT_EDITED', updateDocument),
+  takeEvery('ADD_TO_CONTRACT', addToContract)
 ];
